Add test for ColumnBackend
